refactor: move image routes to blueprint and tidy imports
